fix(build): Resolve Maven compilation and OSGi visibility issues
- Standardized the Maven `groupId` to `org.eclipsetrader` across all `pom.xml` files to resolve parent POM resolution errors.
- Fixed OSGi visibility issues for `javax.xml.bind` by adding it to the `Require-Bundle` list in the `org.eclipsetrader.ui` plugin's manifest.
- Corrected Java 1.6 compatibility issues, including replacing the diamond operator and fixing incorrect casting from `Object` to primitive types.
- Added `target/` directories to `.gitignore` to prevent build artifacts from being tracked.

+++ org.eclipsetrader.jessx/src/org/eclipse/trader/jessx/business/NormalDividend.java
@@ -58,10 +58,10 @@ public class NormalDividend extends Dividend
     public void setParameter(final int i, final Object value) {
         Utils.logger.debug("Setting parameter " + i + ". Value class: " + value.getClass().toString() + ", Value=" + value.toString());
         if (i == 0) {
-            this.mean = (float)value;
+            this.mean = ((Float)value).floatValue();
         }
         else {
-            this.variance = (float)value;
+            this.variance = ((Float)value).floatValue();
         }
     }
